requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.0
Flask-Bcrypt>=1.0.0
Flask-Mail>=0.9.0
mysql-connector-python>=8.0.0
Werkzeug==2.3.7